Don't do a develop build
We expect the preflight to run on a VIP-ready version of the app. There
are a lot of edge cases and weird behaviors with how develop-builds work
and it's hard to account for them.

Instead, let's focus on testing what's important: will a built version
of the app work in production (with a high guarantee)?