fix(build): remove `src` from `.npmignore`
This change removes the `src` folder from `.npmignore` to
fix the issue with publishing.

In 2b0a33e6a32263b6b24a593ab6b8459d4ef634dc, we removed
the build step and, as a consequence, the `dist` folder
became usless: the package was meant to be published
from `src`.

However, because we failed to update `.npmignore`,
we did not publish the source code, and the 1.2.0 release
became broken.

Fixes: PLTFRM-1545